Fix of cloth decommision

+++ backend/app/inventarization/decommissions.py
@@ -40,7 +40,7 @@ async def cloth_decommission(
     cloth: models.Cloth = db.query(models.Cloth).filter(models.Cloth.article == article).one()
 
     new_decommission = models.ClothChanges()
-    new_decommission.area = cloth.width * length
+    new_decommission.area = Decimal(float(cloth.width) * length)
     new_decommission.cloth_article = article
     new_decommission.number = number
     new_decommission.is_income = False
